backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9; platform_system != 'Windows'
pytest>=8.2
pytest-django>=4.8
